+++ CHANGES.rst
@@ -1,15 +1,24 @@
 CHANGES
 =======
 
+0.1.4 (2011.07.08)
+------------------
+
+* Added options for summarySelector, slideSpeed, expandedClass (can be set
+  initially to expand on page load), and ignoredElements (which will not
+  trigger expand/collapse).
+
 0.1.3 (2011.07.06)
 ------------------
 
-* Clicking ``label`` element within ``summary`` does not trigger expand/collapse.
+* Clicking ``label`` element within ``summary`` does not trigger
+  expand/collapse.
 
 0.1.2 (2011.07.06)
 ------------------
 
-* Clicking ``input`` element within ``summary`` does not trigger expand/collapse.
+* Clicking ``input`` element within ``summary`` does not trigger
+  expand/collapse.
 
 0.1.1 (2011.06.25)
 ------------------